[spruce] Normalize host URLs with and without protocol prefix (#180)
## Problem
We want the client to be flexible enough to allow providing
`controllerHostUrl` and `indexHostUrl` with or without the `https://`
protocol. This is useful because the `describeIndex` response includes
host without the prefix, and we're already prepending it there.

## Solution
- Added a new utility method called `normalizeUrl()`.
- Use `normalizeUrl` in `indexOperationsBuilder`,
`vectorOperationsProvider`, and `IndexHostSingleton`. Updating
`indexOperationsBuilder` and `vectorOperationsProvider` covers the cases
where users are providing custom hosts for control or data plane
operations. The `IndexHostSingleton` update replaces the old
`hostWithHttps` utility, and we normalize the host before storing it in
the cache.
- Added unit tests for `normalizeUrl()`.

## Type of Change
- [X] New feature (non-breaking change which adds functionality)

## Test Plan
Verify unit tests and CI pass as expected. The client should continue to
work as expected when resolving data plane URLs for index operations,
etc.

You can verify manually by providing either a `controllerHostUrl` or an
`indexHostUrl` either with or without the prefix:

```
const pinecone = new Pinecone({ apiKey: 'your-api-key', controllerHostUrl: '123-456.svc.io' });

// perform control plane operations, url should be normalized

const index = pinecone.index('test-index', 'index-url-123.svc.pinecone.io');

// perform data plane operations, url should be normalized
```