[spruce] Remove v0 code paths and references (#181)
## Problem
When we launched the node v1.0.0 client, we left in code paths, exports,
and the generated core for the original v0 client. This was to
facilitate ease of migration alongside a migration guide:
`v1-migration.md`.

As we move to release v2.0.0 we'd like to fully deprecate the v0 client.

## Solution

- Remove everything under `/src/v0/` which includes the v0 client code
and generated core.
- Delete `README.v0.md` and `v1-migration.md`.
- Remove references to v0 from `README.md`.
- Remove v0 exports from `src/index.ts`.
- Remove v0 environment requirements and code exports from
`replInit.ts`.
- Remove `isLegacy` from `buildUserAgent`.
- Remove v0 paths and references from config files.

## Type of Change
- [X] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [X] This change requires a documentation update

## Test Plan
Verify CI and tests pass as expected. There should no longer be v0
exports present in the `Pinecone` package.

Run the repl and make sure you're not seeing the legacy export:

```
$ npm run repl

> @pinecone-database/pinecone@1.1.2 repl
> npm run build && node utils/replInit.ts


> @pinecone-database/pinecone@1.1.2 build
> rm -rf dist/ && tsc

INFO Found environment variable PINECONE_API_KEY in .env file
> SUCCESS Pinecone module exports (Pinecone, etc) automatically imported to this repl session.

Run "await init()" to setup client instance using environment variable configs.
```