[spruce] Add bun support in CI (#184)
## Problem
Bun has been a rapidly emerging toolkit for developing, testing,
running, and bundling Javascript and TypeScript. We currently have an
outstanding GH issue related to the bun runtime:
https://github.com/pinecone-io/pinecone-ts-client/issues/141. It would
be nice to add some coverage in CI for running our integration tests
with bun as the runner.

@jhamon kicked this work off a while ago and I wanted to make sure we
got this included.

## Solution
- Update `/github/workflows/testing.yml` and add a new `config` section
to the `integration-tests` matrix. Tests will now be run with `npm` in
node and edge, and `bun` with node for bun versions `1.0.4` -> `1.0.7`.

## Type of Change
- [X] Infrastructure change (CI configs, etc)

## Test Plan
Make sure all the integration tests are 🟢 in CI.