Add new workflow to trigger building and publishing docs, fix release-prod docs step (#185)
## Problem
We pushed a patch to fix up our npm release page. In doing this I
realize the `release-prod` workflow is incorrectly still trying to push
the built client docs to the `gh-pages` branch. It should be moving
things to the `sdk-docs` repo like in `merge`.

## Solution

- Add new `build-and-publish-docs` workflow to allow triggering a fresh
docs build and deploy off main separate from other workflows.
- Update `release-prod` to use this new workflow if the npm release
completes. Also uncommented the test re-use with the boolean flag in
that workflow as it surprised me it immediately pushed to npm without
any other checks.
- Update other workflows to reuse `build-and-publish-docs`.

## Type of Change
- [X] Bug fix (non-breaking change which fixes an issue)
- [X] Infrastructure change (CI configs, etc)

## Test Plan
We'll have to test the release workflow by actually releasing, so give
that a careful review. The docs workflow can be triggered and tested on
its own.