Clean up docstrings (#204)
## Problem
- Leading up to serverless public preview, there was an effort to update
the variable values for the various clients in examples and docs from
`const pinecone = new Pinecone()` to `const client = new Pinecone()`
which didn't make it into the TypeScripts documentation.
- We received feedback from customer success that the current docstring
for `deleteAll()` was possibly confusing to some users, and we wanted to
be more explicit.
- Feedback in this PR:
https://github.com/pinecone-io/pinecone-ts-client/pull/202 around the
strings for `PineconeConnectionError` no longer being accurate since it
references checking client configurations like environment and project
ID which are no longer relevant >v2.0.

## Solution
- Update docstring and message for `PineconeConnectionError` to remove
references to old configuration values. Update the description in all
docstrings where we mark this as thrown.
- Update docstrings for `deleteAll()` to be more explicit around
chaining calls to specific namespaces.
- Update docstrings for a lot of our functions in `pinecone.ts` and
`/data/index.ts`. Most of this is stylistically aligning things, moving
from using `pinecone` as a variable to `client`, and adding some missing
`@returns` and `@throws` notation.
- Update `README.md` to clean up some of our code examples.

## Type of Change
- [X] Non-code change (docs, etc)

## Test Plan
Re-build docs locally and open the HTML to make sure things look as
expected:

```
npm run docs:build
```

This will generate a `/docs` folder at the root of the repo, open the
HTML file in here.