Update CI to test newer versions of Node and test fewer versions of Bun (#228)
## Problem

We currently only test Node version `16.x`, but there are much newer
versions with upgraded capabilities. To ensure full coverage, we want to
add in `18.x` and `20.x`. This PR does that.

Note: this will increase the time it takes for our CI pipeline to
complete, but we have discussed, and the increased latency is worth it
to cover more modern versions of Node.

We have also **decreased** the number of `bun` versions tested since
they have released a major new version `1.1.x`. We originally had many
versions out of an abundance of caution leading up to releasing Pinecone
serverless. Now that there have been major releases of `bun` cut more
recently, it's my opinion that we only really need to test the
following:
- `1.0.0`, which is basically when they GAed `bun` (and we've always
tested this initial major release)
- `1.0.36`, the last minor release before `1.1.0`.
- `1.1.11`, the most recent version of `bun`, which works with Windows
and fixes some bugs found in the initial `1.1.0` release.

Our CI pipeline still takes a while, but ultimately I think this is
better test coverage, and we can do a quick follow-up PR with some hacks
to decrease the overall time at some point in the near future.


## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [x] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)



---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1207545518662996