Add Inference API /embed endpoint to July RC branch (#230)
Add `/embed` endpoint to July RC branch.

**Note:**
- This endpoint is only available on the July RC branch (`2024-07-rc`).
In order to use the new `/embed` endpoint, users will need to clone this
repository, fetch the RC branch, build the project, and play around with
the endpoint that way.
- As highlighted in the README, the `/embed` endpoint code only works
with Node `>=18`. **Failed integration tests for Node `<18` are expected
in this PR**, excluding the run with the `edge` runtime (this runtime
provides a `global.fetch` that overrides the ponyfill, so the root issue
with Node versions `<18` and how they implement `fetch` is not a
problem).

**Included in this PR are:**
- New generated core files
- A new `inference` directory, which contains the code users should use
when interacting with this endpoint
- A new `models` directory that contains wrappers around generated data
structures. This mimics the organization we have currently in [the
Python
client](https://github.com/pinecone-io/pinecone-python-client/tree/main/pinecone/models).
- Integration and unit tests

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [x] Non-code change (docs, etc)
- [ ] None of the above: (explain here)