Add workflow and action files for RC code (#231)
## Problem

We would like to publish RC (release candidate) code to `npm`. In order
to do that, we need new github workflow and action files. This PR adds
those files, so that we can use the RC branch (`2024-07-rc`) to push to
`npm` for users.
"
Since versioning RCs is a bit of an antipattern, these workflows do
_not_ contain versioning commands. The "version" is controlled by this
command, which just appends `.rc`: `run: npm version prerelease
--preid=rc --no-git-tag-version`.

TBD on whether we need Github tags, either. Would love reviewers' input
on that (I'm thinking we don't).


## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [x] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)