Add new `CreateIndexSpec` types, regenerate from latest OpenAPI spec (#237)
## Problem
There was an internal discussion recently involving `shards` and
`replicas` fields in `PodSpec` which were not marked as required. This
was originally done to address build issues with generated code in
TypeScript, but is an inaccurate representation of our spec.

`PodSpec` was addressed in this PR, and we need to regenerate the
TypeScript code and fix: https://github.com/pinecone-io/apis/pull/102

## Solution
- Regenerate `pinecone-ts-client-fetch` from the latest OpenAPI spec
using `npm run generate:openapi`.
- Add new `TypeScript` types so we can maintain the flexibility of some
of these optional parameters inside `IndexSpec` when creating an index:
`CreateIndexSpec`, `CreateIndexServerlessSpec`, and
`CreateIndexPodSpec`. Previously, we were relying on the generated types
for everything in create, but there are specific nuances to what's
required and what's not, and this seems like the most reasonable way of
addressing it at the client level.
- Small fix in `codegen/build-oas.sh` - we need to run `git checkout
main` in the submodule when updating.
  
You can ignore changes under `/pinecone-generated-ts-fetch/*` which is
where all the code generated from OpenAPI lives.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
Run unit and integration tests to validate the `createIndex` operation
is still working as expected. We're exporting new types from the package
and other than that, the changes here shouldn't be apparent from a UX
perspective.