Remove support for Node `<18.0.0` (#239)
## Problem
When implementing inference @aulorbe ran into issues related to large
response payloads causing timeouts in integration tests. She tracked
this down to being related to the specific version of Node being used to
run the integration tests. Specifically, this was a problem with
versions `<=16.0.0`.

We currently support Node `>14.0.0`, however `v20.0.0` is the current
LTS version. `v14.0.0` was EOL April 30, 2023, and `v16.0.0` in
September 11th, 2023. Details in this Slack thread:
https://pinecone-io.slack.com/archives/C05MYEX7T2L/p1718202858574909?thread_ts=1718135375.511479&cid=C05MYEX7T2L

## Solution
Given the issues we've run into testing on older versions of Node, and
these versions being fully EOL'd, we should bump our engine support to
`v18.0.0` for now. Details in Slack:
https://pinecone-io.slack.com/archives/C05MYEX7T2L/p1718202858574909?thread_ts=1718135375.511479&cid=C05MYEX7T2L

- Update `package.json` to support Node `>=18.0.0`.
- Remove `v16.0.0` from integration testing matrix.

## Type of Change
- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
`npm run build`
`npm run test:integration`

Validate we've got green across the board for inference integration
tests in CI.