Implement Deletion Protection (#238)
## Problem
`deletionProtection` is part of the control plane spec, but we have not
implemented it in the TypeScript SDK.

## Solution
- Update `configureIndex` schema and `options` type to
`ConfigureIndexRequest` to properly support the new object structure for
making requests to configure an index. We needed `deletionProtection` at
the top level, and the new nesting for `spec: { pod : { ... }}`.
- Update `createIndex` schema to validate optional `deletionProtection`.
- Refactor `configureIndex` unit / validation tests.
- Refactor / enable `configureIndex` integration tests - these had been
disabled during the serverless launch era and I never re-wrote them.
Better late than never.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [X] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

- Test that creating and index without passing `deletionProtection`
defaults to `disabled`.
- Confirm that you can configure both pod and serverless index
`deletionProtection` options.
- Confirm that trying to delete an index with `deletionProtection:
'enabled'` throws an error.
- Validate we're seeing the `deletionProtection` field in operations
such as `listIndexes` and `describeIndex`.