Add console warning when using SDK in browser context (#240)
## Problem
We recently had an issue logged around problems using the Pinecone SDK
from a browser context because we're applying a user-agent to outgoing
HTTP requests, and certain browsers don't play nicely with that.

Ultimately, people should not be using the SDK from a browser as it's a
large security risk, and exposes their API key if deployed to
production. We'd like to add a warning when we detect the SDK is being
used from the browser.

## Solution
- Add `isBrowser` function to `utils/environment.ts`.
- Add private helper `_checkForBrowser()` to `Pinecone` which calls
`isBrowser` and calls `console.warn` if true. This is run once when
constructing the `Pinecone` class.
- Add a unit test to validate the behavior when `window` is present in
the global object.
- Use the generated API version default `X_PINECONE_API_VERSION` in the
`inferenceOperationsBuilder`.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
You'll need to run the SDK from within a browser, which would take some
setup. The unit test should be good enough to cover the behavior here.