Move secrets only to workflow and not action file (#259)
## Problem

The end to end testing workflow was getting errors related to secrets,
e.g.:

>
/home/runner/work/pinecone-ts-client/pinecone-ts-client/./.github/actions/e2e-testing/edge/action.yml
(Line: 20, Col: 25): Unrecognized named-value: 'secrets'. Located at
position 1 within expression: secrets.VERCEL_TOKEN

After reading this thread
https://github.com/orgs/community/discussions/27054#discussioncomment-3254450
it seems that `secrets` might only be available in workflows and not
actions. So, I've moved `secrets` to only be in the workflow file and
the references to them to be in `inputs` in the `action.yml` file.


## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [x] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)