Change language around records --> vectors where appropriate (#288)
We believe it's more accurate to refer to dense-vector objects as
"vectors" (vs "records") when we are talking about vectors we'd like to
upsert into an index, etc. This is because they are primarily (first and
foremost) _vectors_, with IDs, metadata, etc.

Change instances of "record(s)" to "vector(s)" where appropriate.

Note: I opted _not_ to change instances of "record(s)" in cases where we
are talking about the entire object as a whole (as in, "Update a
record", since you could be updating the metadata, the vector values,
etc.).

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [x] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1208381566824289