Add integration testing w/external NextJS app (Edge runtime) (#304)
## Problem

We do not currently have a good way to test how our client behaves end
to end in different environments. One of the chief problems our users
have brought to our attention in the past is that some functionalities
do not work in certain runtimes, e.g. Edge or Bun, and/or with certain
frameworks, e.g. [NextJS](https://nextjs.org/docs).

## Solution

Build an external application in a runtime and framework known to have
caused problems in the past and test our client from the end-user's
perspective.

This PR introduces and end-to-end test suite that interacts with an
external application written using the NextJS framework and the
[Edge](https://vercel.com/docs/functions/edge-middleware/edge-runtime#edge-runtime)
runtime.

Note: the app is automagically run in `Edge` by way of it using
[middleware](https://nextjs.org/docs/app/building-your-application/routing/middleware)
(`middleware.ts`), and the [`Headers()`
API](https://developer.mozilla.org/en-US/docs/Web/API/Headers).

## Overview of changes
- We now have [a public
repo](https://github.com/pinecone-io/ts-client-e2e-tests) that contains
a super simple application that builds a Pinecone serverless index,
seeds it with data, and queries that data. We plan to add more
operations in the near future; this is just a start.
- This sample application creates an API endpoint that our client repo
can send a POST request to and assert on the response. If the endpoint
fails to deliver the expected response, we know that something is wrong
with the most recent changes we are proposing to introduce in the
client.
- This PR enables 2 types of interaction with the sample application:
- Local end-to-end tests: for local runs, there is a new bash script
that devs can execute. This will spin up the application on their
`localhost:3000`.
- CI end-to-end tests: for CI runs, we will now run this test
automatically by way of adding it the `testing.yml` file that is run for
each push to an open PR. In CI, the Github workflow and action hit the
application's Vercel endpoint and assert on its response.
- There is a new CONTRIBUTING.md file with some info on this new test
suite + the other existing ones, as well as a new README in the
`external-app` dir itself with more in depth information.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] This change requires a documentation update
- [x] Infrastructure change (CI configs, etc)
- [x] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

If reviewers can pull down the code in this branch and try to run the
tests locally, that'd be great.

## To Dos
I'll add a README and a CONTRIBUTING file to [the external
app](https://github.com/pinecone-io/ts-client-e2e-tests)

---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1208129226784956