Add Retries to Upsert operation (#310)
## Problem

We do not have a way to retry common data plane operations, such as
`upsert`. This is a feature that's often requested by the community and
needed generally for a more mature client.

## Solution

Add a new class called `RetryOnServerFailure` that handles retries. As a
POC, integrate that class into the `upsert` operation.

## Overview of changes
- Users can now pass a `maxRetries` property to the
`PineconeConfiguration` object if they want to tweak how many times
retryable ops retry (default is 3, and only on `upsert` for now).
- New `RetryOnServerFailure` class: takes in an asynchronous operation
(e.g. `upsert`) and retries it if the server respond with either of 2
new errors: a `PineconeUnavailableError` (`503` error) or a
`PineconeInternalServerError` (`500` error).
- This new `RetryOnServerFailure` class retries `3` times as a default
with an exponential backoff and a jitter factor
- If `maxRetries` exceeds `10`, we now throw an error to avoid
overloading the server

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [x] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

New unit and integration tests pass. 

---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1206468139133724