Add newly-generated code for Assistant data, control, and eval planes to Assistant feat. branch (#315)
For an unknown-as-of-right-now reason, the Typescript `fetch` generator
has been failing non-deterministically to generate code. We successfully
got it to generate on a new branch, so this PR simply merges the
generated code to the existing Assistant feature branch.

I will hunt around online to see if I can find any open Github issues
that might reveal what's going on w/the `fetch` generator...

[Note the only diff btwn these 2 branches is the generated code,
isolating the generation issues to the generator itself, not to files in
our client repo.]