Remove `EmbeddingsList` custom class (#330)
## Problem
The `EmbeddingsList` class was extremely complicated, and was written
only to truncate lengthy output to the console for demonstration
purposes.

The class was incorrectly extending `Array<>`, and mangling the data we
were getting back from `embed` calls that included sparse details.

## Solution
Remove the custom EmbeddingsList wrapper and tests and use the interface
from the generated core. This means we'll actually get back what we're
expecting, and we remove a complex, unwieldy chunk of code.

| Before  | After |
| ------------- | ------------- |
|![Screenshot 2025-02-19 at 8 14
54 PM](https://github.com/user-attachments/assets/4e868225-524c-47d7-8c7f-1fb9a12f76d5)|![Screenshot
2025-02-19 at 8 10
22 PM](https://github.com/user-attachments/assets/a5ca19b8-16e0-4700-86f8-fbb832dfcb01)|




## Type of Change
- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan